perception: extract plugin object-recognition
This is the second part of the split-up of tabletop-objects; this plugin
does the object recognition. It reads the object pointcloud (the points
above the table excluding the table), which is published by the
tabletop-recognition plugin. It clusters the pointcloud and publishes
all centroids to the blackboard and all object clusters as (seperate)
pointclouds.

Note that no object tracking or cylinder fitting happens here.