obj-tracking: check old centroids again if an ID was wrongly assigned
If an ID was wrongly assigned (i.e. the new object is too far away from
the old object), we used to simply drop the object. In certain
situations, this leads to a lot of dropped objects.
Instead of dropping the object, check the old centroids for a match.
If no old object is found, add the object as new object.