perception: always set visibility_history to 0 if plugin is disabled
Until now, all the perception plugins didn't change visibility when they
were disabled. This can lead to problems with other components which
enable the plugin and immediately read (outdated) blackboard data.
Therefore, always set the visibility_history to 0 if the plugin is
disabled.  Note that the positions are not updated (and thus still
accessible through the blackboard interface).