openprs: Sleep Wait Before Connecting To PRSServer
The problem in this is that if the connection is refused
once (For example, in case oprs-server didn't have enough
time to start), the whole connection operation is aborted
and the processes running the oprs-server and mp are also
closed.
It will only succeed if the server is guaranteed to be up
by the time we try to connect. Another solution would be
to busy wait till the server is up.

Bottom line, It will attempt to establish the connection only once!