openprs-agent: included meta-ops for Goal-Intending
Included meta-intended-goal.inc. Contains Ops to allow directly
intending a new intention which has to achieve a particular goal.

This is a different method for intending goals as opposed to the default
"Applicable-Op-Intending".

For example you can build a goal and intended with ops like
INTENDED-OP $goal or INTENDED-ALL-OPS-AS-ROOT $goal-list