pddl-robot-memory: retrieve all data in a single aggregation pipeline
This uses the facet feature of MongoDB aggregation to get all data in a
single query in order to ensure consistency.

This is currently a test version which replaces the previous query code.
It has the drawback that the all results together must fit into a single
document.

For a final version this will need to be made optional. This is only an
intermediate commit not to be merged.