clips-executive: support actions with the same neg and pos effect
If an action has the same effect both as positive and negative effect,
assume that we always want to have the positive effect. To achieve this,
check if there is any positive effect with the same predicate and same
parameters and only apply the negative effect if no such positive effect
exists.