plexil: better handling of setting an unknown global state
If the type of the passed value is UNKNOWN, do not put it into the
values map as we wouldn't gain anything (the value would be unknown
either way), but we cannot detect inconsistencies later. Additionall use
Fawkes debug log output for less intrusive information when adding
values. Fix a glitch that prevented proper creation of multiple
parametrized states with the same name.