plexil: support timer re-scheduling
Plexil might set a timer while it already had configured one, e.g., if a
newly running action requires an earlier trigger. So far, we did not
support this. Now, the timer support thread can re-schedule for new
timer events. Additionally, we check if the intended trigger time has
already passed and if it has, we trigger the timer right away.